Generate minimal bodied tag.
Implement the generation of a minimal bodied tag. Added a `generating`
variable to the `rewrite` and directive handler signatures. It indicates
that a body has just be generated, either by first generating the
document using the `generate` function, or else by introducing a
directive body to an as a result of an `if` or `each` directive in a
call to the `regenerate` function. This implementation should support
the default block bodies defined in the templates.

Closes #72.