Layouts.
Implement layouts with named blocks.

Leaning toward using ECMA 5 and requesting that IE 8 users shim. It is
the right thing to do, to move everyone over to the new standards. IE 8
is the new IE 6, a laggard that requires special preprocessing, we
always have one Microsoft browser we're molly-coddling, so let it now be
IE 8, but don't let the molly-coddling ruin the code.

I realy only want to use `forEach` and `filter`, which might reduce to just
`filter`, and then there is only one function to shim. I'm going to work
with `forEach` and `filter`, then later I'll see how much the size
explodes when I shim `filter` or replace `filter` with `for` loops.

Closes #63.