Tag attributes, `$context` variable.
Expose tag attributes in the context of the tag directives. Directives
defined within a tag can obtain the attribute values of the attributes
defined on the tag in an `$attibutes` object. Only unqualified
attributes are added to the `$attributes` object.

Additionally, the context is now exposed through the `$context` object,
which may be shortened to `$` in a future commit.

Closes #80.