Strip `xmlns` based directives.
Still no intention of supporting XML, but I am leaving `xmlns`
directives that are not related to Stencil in the documents.

No intention of supporting XML because Stencil is a DOM manipulation
library, a DOM templating library, not a string templating library, and
not an XML templating library. It is indented for web applications,
building documents to be viewed in a web browser. Stencil is not an
HTML5 generator. HTML5 is the chosen standard for serialization of
Stencil documents.

Closes #64.