Sub-attributes and frame management.
Implemented sub attributes, the ability of a sub tag to reference the
attributes of the tag that defined it.

With this commit, we begin to do some frame management, maintaining a
state that is consistent across invocations of sub-tags. In a subsequent
commit we will ensure that an included tag does not see the context of
the caller, that it has it's own clean context inheriting the context of
the `include` element that defined the tags.

Closes #86.