Split `resolver` into `modules` and `resources`.
Module resolution is a different beastie. Not in the mood to have a go
at sounding intelligent about it at the time of this commit, so I'm
going to say that the module mechanism is used to load modules, while
`readFile` is used to load XML and JSON. The two are significantly
different.

One the client side, I want to use AMD to load modules, but I'm going to
want to stick with XHR (and some sort of JSONP implementation) to load
JSON and XML.

I don't want to start going down a path of creating a grand unified
resource abstraction. XHR works well. Loading scripts works poorly, so
I'm going to defer to AMD for that. Down the road, I might want to offer
alternatives to AMD to load modules.

Actually, this is all dubious. Stencils include libraries, so we need to
resolve paths relative to the Stencil, so how can we read Stencils from
HTTP and modules using `require`?

Closes #96.