Separate resolvers for each resource type.
Decided that it was easiest for now to have a separate resolver for each
resource type; `javascript`, `xml`, and `json`, where the base is
determined through a constructor function for the resolver. Stencil
resolves all URLs relative to a psudeo-base, the resolvers prepend
whatever prefix is necessary to resolve resource.

As I'm writing this, it occurs to me that JSON, while key to Stencil, is
nothing that Stencil needs to resolve for itself. That ought to take
place in the scaffolding for the stencils.

Closes #97.

File                             Raw         Min         Zip     Min/Zip
index.js                 28104/27.45  8908/08.70  7096/06.93  3397/03.32