Implement `recurse` directive.
I didn't want to over-think it, so I simply added a new parameter to the descent
of the template; `named` which has all the named attributes. It uses `extend` to
include them in a frame, so that the named `each` directives appear in
scope like the context. The `named` maps the name of a named `each`
directive to its directive object.

Closes #102.