Implement a minimal Stencil parser.
Start over with the Stencil parser for the curly braced version of the
Stencil language. This parser is going to be a brute; non-incremental,
recursive-descent, look-ahead as far as the eye see, with regular
expressions run against substrings. A brute.

But, it should be pretty easy to write and maintain. It doesn't have to
parse HTML, just templates, which ought not be as wild and crazy as the
HTML found in the wild. Plus, the parser is run only once, by the
server, to convert the Stencil language into the behind-the-scenes XML
format.

Closes #175.